chore: rename featureextack and preloginrequired
