+++ src/token/stream-parser.js
@@ -9,7 +9,7 @@ tokenParsers[TYPE.DONEPROC] = require('./done-token-parser').doneProcParser;
 tokenParsers[TYPE.ENVCHANGE] = require('./env-change-token-parser');
 tokenParsers[TYPE.ERROR] = require('./infoerror-token-parser').errorParser;
 tokenParsers[TYPE.FEDAUTHINFO] = require('./fedauth-info-parser');
-tokenParsers[TYPE.FEATUREEXTACK] = require('./feature-ext-parser');
+tokenParsers[TYPE.FEATUREEXTACK] = require('./feature-ext-ack-parser');
 tokenParsers[TYPE.INFO] = require('./infoerror-token-parser').infoParser;
 tokenParsers[TYPE.LOGINACK] = require('./loginack-token-parser');
 tokenParsers[TYPE.ORDER] = require('./order-token-parser');
